move timeseries_api package to api_clients.timeseries
