+++ src/enlyze/api_clients/timeseries/models.py
